requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
langchain==0.2.16
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3
langchain-community==0.2.16
llama-index==0.10.5
pinecone-client==5.0.0
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
mlflow==2.14.0
pydantic-settings==2.2.0
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
